requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
python-multipart